feat(update lm flow): replace anchor with spl (#7)
